# animals-service/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Animals Service API",
    "description": "Gestión de expedientes y movimientos",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Registrar nuevo usuario",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre":     { "type": "string", "example": "Juan" },
                  "apellido_p": { "type": "string", "example": "Pérez" },
                  "apellido_m": { "type": "string", "example": "García" },
                  "email":      { "type": "string", "example": "juan@refugio.com" },
                  "contrasena": { "type": "string", "example": "******" },
                  "activo":     { "type": "boolean", "example": true },
                  "rol_id":     { "type": "string", "format": "uuid" },
                  "refugio_id": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuario registrado exitosamente" },
          "409": { "description": "Email ya registrado" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email":      { "type": "string", "example": "admin@refugio.com" },
                  "contrasena": { "type": "string", "example": "********" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login exitoso, retorna access_token" },
          "401": { "description": "Credenciales inválidas" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Autenticación"],
        "summary": "Obtener perfil del usuario autenticado",
        "responses": {
          "200": { "description": "Perfil del usuario" },
          "401": { "description": "No autenticado" }
        }
      }
    },
...
    "/animals": {
      "post": {
        "tags": ["Animales"],
        "summary": "Crear un animal",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["nombre", "estado", "especie", "raza", "edad", "peso", "sexo", "tamano", "lugar", "descripcion", "refugio_id", "usuario_id"],
                "properties": {
                  "nombre":                 { "type": "string", "example": "Firulais" },
                  "estado":                 { "type": "string", "enum": ["adoptado", "adopcion", "recuperacion", "defuncion", "extraviado"]},
                  "especie":                { "type": "string", "example": "Perro" },
                  "raza":                   { "type": "string", "example": "Labrador" },
                  "edad":                   { "type": "integer", "example": 3 },
                  "peso":                   { "type": "number", "example": 12.5 },
                  "sexo":                   { "type": "string", "enum": ["Macho", "Hembra"] },
                  "tamano":                 { "type": "string", "enum": ["miniatura", "pequeno", "mediano", "grande", "gigante"]},
                  "enfermedad_no_tratable": { "type": "boolean", "example": false },
                  "discapacidad":           { "type": "boolean", "example": false },
                  "es_agresivo":            { "type": "boolean", "example": false },
                  "lugar":                  { "type": "string", "example": "Jardín" },
                  "descripcion":            { "type": "string", "example": "Animal tranquilo" },
                  "refugio_id":             { "type": "string", "format": "uuid" },
                  "usuario_id":             { "type": "string", "format": "uuid" },
                  "imagen":                 { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Animal creado" }
        }
      }
    },
    "/animals/refugio/{refugio_id}": {
      "get": {
        "tags": ["Animales"],
        "summary": "Obtener animales por refugio",
        "parameters": [
          { "name": "refugio_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Lista de animales" },
          "404": { "description": "Refugio no encontrado" }
        }
      }
    },
    "/animals/{id}": {
      "get": {
        "tags": ["Animales"],
        "summary": "Obtener un animal por ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Animal encontrado" },
          "404": { "description": "Animal no encontrado" }
        }
      },
      "patch": {
        "tags": ["Animales"],
        "summary": "Actualizar un animal",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre":                 { "type": "string" },
                  "estado":                 { "type": "string", "enum": ["adoptado", "adopcion", "recuperacion", "defuncion", "extraviado"] },
                  "especie":                { "type": "string" },
                  "raza":                   { "type": "string" },
                  "edad":                   { "type": "integer" },
                  "peso":                   { "type": "number" },
                  "sexo":                   { "type": "string", "enum": ["Macho", "Hembra"] },
                  "tamano":                 { "type": "string", "enum": ["miniatura", "pequeno", "mediano", "grande", "gigante"] },
                  "enfermedad_no_tratable": { "type": "boolean" },
                  "discapacidad":           { "type": "boolean" },
                  "es_agresivo":            { "type": "boolean" },
                  "lugar":                  { "type": "string" },
                  "descripcion":            { "type": "string" },
                  "imagen":                 { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Animal actualizado" },
          "404": { "description": "Animal no encontrado" }
        }
      },
      "delete": {
        "tags": ["Animales"],
        "summary": "Eliminar un animal",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Animal eliminado" },
          "404": { "description": "Animal no encontrado" }
        }
      }
    },
...
    "/movements/animal/{animal_id}": {
      "get": {
        "tags": ["Movimientos"],
        "summary": "Obtener movimientos por animal",
        "parameters": [
          { "name": "animal_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Lista de movimientos del animal" },
          "404": { "description": "Animal no encontrado" }
        }
      }
    },
...
  }
}